backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
# psycopg2-binary==2.9.9  # для PostgreSQL (закомментировано, используем SQLite)

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator==2.1.0

# CORS
python-dotenv==1.0.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
